Update compilation issues when building  64-bit slices

+++ Classes/YAJLGen.m
@@ -122,16 +122,16 @@ NSString *const YAJLGenInvalidObjectException = @"YAJLGenInvalidObjectException"
     yajl_gen_null(gen_);
   } else {
     NSString *s = [number stringValue];
-    unsigned int length = [s lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
+    NSUInteger length = [s lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
     const char *c = [s UTF8String];
-    yajl_gen_number(gen_, c, length);
+    yajl_gen_number(gen_, c, (unsigned int) length);
   }
 }
 
 - (void)string:(NSString *)s {
-  unsigned int length = [s lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
+  NSUInteger length = [s lengthOfBytesUsingEncoding:NSUTF8StringEncoding];
   const unsigned char *c = (const unsigned char *)[s UTF8String]; 
-  yajl_gen_string(gen_, c, length);
+  yajl_gen_string(gen_, c, (unsigned int) length);
 }
 
 - (void)startDictionary {
